run `forge fmt`

+++ test/pumps/PumpHelpers.sol
@@ -15,8 +15,7 @@ function to18(bytes16 a) pure returns (uint result) {
     return uint(ABDKMathQuad.to128x128(a)) * MAX_E18 / MAX_128;
 }
 
-
-function simCapReserve50Percent(uint256 lastReserve, uint256 reserve, uint blocks) view returns (uint256 cappedReserve) {
+function simCapReserve50Percent(uint lastReserve, uint reserve, uint blocks) view returns (uint cappedReserve) {
     uint limitReserve = lastReserve * 1e18;
 
     uint multiplier = lastReserve < reserve ? 1.5e6 : 0.5e6;
@@ -38,16 +37,17 @@ function simCapReserve50Percent(uint256 lastReserve, uint256 reserve, uint block
     console.log("lastReserve", lastReserve);
     console.log("reserve", reserve);
 
-    cappedReserve =
-        (lastReserve < reserve && limitReserve < reserve) ||
-        (lastReserve > reserve && limitReserve > reserve) ?
-        limitReserve : reserve;
+    cappedReserve = (lastReserve < reserve && limitReserve < reserve)
+        || (lastReserve > reserve && limitReserve > reserve) ? limitReserve : reserve;
 }
 
-function generateRandomUpdate(uint n, bytes32 seed) pure returns (uint256[] memory balances, uint40 timeIncrease, bytes32 newSeed) {
+function generateRandomUpdate(
+    uint n,
+    bytes32 seed
+) pure returns (uint[] memory balances, uint40 timeIncrease, bytes32 newSeed) {
     balances = new uint[](n);
     seed = stepSeed(seed);
-    timeIncrease = uint40(uint256(seed)) % 50_000_000;
+    timeIncrease = uint40(uint(seed)) % 50_000_000;
     for (uint i; i < n; ++i) {
         seed = stepSeed(seed);
         balances[i] = uint(uint128(uint(seed))); // case to uint128
